[refactor] 자동이체 코드 품질 개선
